+++ tools/integration_tests/read_large_files/read_one_large_file_sequentially_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package operations
+package read_large_files
 
 import (
 	"bytes"
