+++ tools/integration_tests/read_large_files/testdata/write_content_fix_size.sh
